ai-service/requirements.txt
# fastapi==0.111.0
# uvicorn[standard]==0.30.1
# google-generativeai==0.8.3
# python-dotenv==1.0.1
# pydantic==2.7.1
# httpx==0.27.0
# requests==2.32.3
# beautifulsoup4==4.12.3
# playwright==1.44.0
# sentence-transformers==3.0.1
# numpy
# gunicorn==22.0.0


fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo==4.7.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
python-dotenv==1.0.1
google-generativeai==0.7.2
beautifulsoup4==4.12.3
requests==2.32.2
playwright==1.44.0
authlib==1.3.1
itsdangerous==2.2.0
starlette==0.37.2
aiofiles==23.2.1
uuid==1.30